+++ DepthStencilBuffer.h
@@ -9,7 +9,7 @@ private:
 public:
 	DescriptorHeap heap;
 	using Hresultable::Hresultable;
-	DepthStencilBuffer(ID3D12Device10* device, UINT width, UINT height);
-	void Resize(ID3D12Device10* device, UINT w, UINT h);
+	DepthStencilBuffer(ID3D12Device10* device, uint16_t width, uint16_t height);
+	void Resize(ID3D12Device10* device, uint16_t w, uint16_t h);
 };
