+++ SwapChain.cpp
@@ -1,6 +1,6 @@
 #include "SwapChain.h"
 
-SwapChain::SwapChain(IDXGIFactory7* factory, ID3D12CommandQueue* queue, HWND hwnd, UINT width, UINT height)
+SwapChain::SwapChain(IDXGIFactory7* factory, ID3D12CommandQueue* queue, HWND hwnd, uint16_t width, uint16_t height)
 {
 	descriptor.BufferCount = 3;
 	descriptor.BufferUsage = DXGI_USAGE_RENDER_TARGET_OUTPUT;
@@ -9,8 +9,6 @@ SwapChain::SwapChain(IDXGIFactory7* factory, ID3D12CommandQueue* queue, HWND hwn
 	descriptor.Format = DXGI_FORMAT_R8G8B8A8_UNORM;
 	descriptor.Height = height;
 	descriptor.Width = width;
-	//descriptor.Height = height;
-	//descriptor.Width = width;
 	hr = factory->CreateSwapChainForHwnd(queue, hwnd, &descriptor, NULL, NULL, &tmp);
 	if (Ok()) {
 		hr = tmp->QueryInterface<IDXGISwapChain4>(&ptr);
